# docs/swagger.yaml
basePath: /api/v2
...
host: localhost:8080
info:
  contact: {}
  description: Enterprise gateway for Autonomi decentralized storage. Provides file
    upload/download, user management, API tokens, webhooks, and admin controls.
  title: Indelible API
  version: "2.0"
paths:
...
  /admin/logs/audit:
    get:
      description: Return audit log entries with optional filtering (permanent, never
        deleted)
      parameters:
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      - description: Max results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Query audit logs
      tags:
      - 'Admin: Logs'
  /admin/logs/config:
    get:
      description: |-
        Return system log entries with optional filtering (retention-managed)
        Return config_audit entries showing old/new value, actor, IP, and UA for every settings change. Permanent (never purged).
      parameters:
      - description: Filter by log level
        in: query
        name: level
        type: string
      - description: Filter by component
        in: query
        name: component
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      - description: Max results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by setting key
        in: query
        name: setting_key
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      - description: Max results (default 100, max 500)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Query config-change audit log
      tags:
      - 'Admin: Logs'
      - 'Admin: Logs'
  /admin/logs/system:
    get:
      responses: {}
      security:
      - BearerAuth: []
  /admin/logs/user:
    get:
      description: Return user activity log entries with optional filtering
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      - description: Max results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Query user activity logs
      tags:
      - 'Admin: Logs'
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT or API token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"